+++ playground/basic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # basic
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [14312a2]
+  - @leanweb/fullstack@0.3.1
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [9e0ecbe]
+  - @leanweb/fullstack@0.3.0
+
 ## 1.0.10
 
 ### Patch Changes